Remove truncation which could break huge exception handlers. (2). (thanks janosz)

+++ src/org/benf/cfr/reader/entities/exceptions/ExceptionAggregator.java
@@ -242,7 +242,7 @@ public class ExceptionAggregator {
             int indexFrom = exceptionTableEntry.getBytecodeIndexFrom();
             if (indexFrom < handlerIndex &&
                     indexTo >= handlerIndex) {
-                exceptionTableEntry = exceptionTableEntry.copyWithRange((short) indexFrom, (short) handlerIndex);
+                exceptionTableEntry = exceptionTableEntry.copyWithRange(indexFrom, handlerIndex);
             }
             extended.add(exceptionTableEntry);
         }
